refactor: move register_coders!
