+++ sdks/rust/src/coders/register_coders/mod.rs
@@ -1,0 +1,29 @@
+/// Must be called in outside of the main() function.
+///
+/// TODO example
+#[macro_export]
+macro_rules! register_coders {
+    ($($coder:ident),*) => {
+        fn encode_from_urn(urn: &str, elem: &dyn $crate::elem_types::ElemType, writer: &mut dyn std::io::Write, context: &$crate::coders::Context) -> Result<usize, std::io::Error> {
+            match urn {
+                $($coder::URN => $coder.encode(elem, writer, context),)*
+                _ => panic!("unknown urn"),
+            }
+        }
+
+        fn decode_from_urn(urn: &str, reader: &mut dyn std::io::Read, context: &$crate::coders::Context) -> Result<Box<dyn $crate::elem_types::ElemType>, std::io::Error> {
+            match urn {
+                $($coder::URN => $coder.decode(reader, context),)*
+                _ => panic!("unknown urn"),
+            }
+        }
+
+        #[ctor::ctor]
+        fn init_coders_from_urn() {
+            $crate::coders::CODERS_FROM_URN.set($crate::coders::CodersFromUrn {
+                enc: encode_from_urn,
+                dec: decode_from_urn,
+            }).unwrap();
+        }
+    }
+}
